+++ source/include/Ostap/MatrixUtils.h
@@ -1771,6 +1771,197 @@ namespace Ostap
       return true ;
     }
     // ========================================================================
+    
+    // ========================================================================
+    /// Is this matrix symmetric ?
+    template <class T, unsigned int D1, unsigned int D2, class R>
+    inline bool  symmetric
+    ( const ROOT::Math::SMatrix<T,D1,D2,R>& /* mtrx */ ) 
+    { return false ; }
+    // ========================================================================
+    /// Is this matrix symmetric ?
+    template <class T, unsigned int D, class R>
+    inline bool  symmetric
+    ( const ROOT::Math::SMatrix<T,D,D,R>& mtrx) 
+    {
+      const Ostap::Math::Equal_To<T> equal ;
+      for ( unsigned i = 0 ; i < D ; ++i )
+      { for ( unsigned j = 0 ; j < i ; ++j )
+	{ if ( !equal ( mtrx ( i , j ) , mtrx ( j , i ) ) ) { return false ; } } }
+      return true ;
+    }
+    // =======================================================================
+    /// Is this matrix symmetric ?
+    template <class T, unsigned int D>
+    inline bool  symmetric
+    ( const ROOT::Math::SMatrix<T,D,D,ROOT::Math::MatRepSym<T,D>> & /* mtrx */ )
+    { return true ; }
+    // =======================================================================
+
+    // =======================================================================
+    /// Can this matrix be symmetric & positive-definite ?
+    template <class T, unsigned int D1, unsigned int D2, class R>
+    inline bool symmetric_positive_definite 
+    ( const ROOT::Math::SMatrix<T,D1,D2,R> & /* mtrx */ )
+    { return false ; }
+    // ========================================================================
+    /** Can this matrix be symmetric & positive-definite ?
+     *  - Diagonal elements are finite and positive
+     *  - Symmetric
+     *  - Off-diagonal elements are finite
+     *  - have CholeskyDecomposition
+     */
+    template <class T, unsigned int D, class R>
+    inline bool symmetric_positive_definite 
+    ( const ROOT::Math::SMatrix<T,D,D,R> & mtrx )
+    {
+      //
+      const Ostap::Math::Zero<T>     zero  ;
+      const Ostap::Math::Equal_To<T> equal ;
+      //
+      for ( unsigned int i = 0 ; i < D ; ++i )
+      {
+	/// diagonal elements must be positive! 
+	const T mii = mtrx ( i , i ) ;
+	if ( !std::isfinite ( mii ) || mii <= 0 || zero ( mii ) ) { return false ; } // positive ? 
+	for ( unsigned j = 0 ; j < i ; ++j )
+	{
+	  //
+	  const T mij = mtrx ( i , j ) ;
+	  if ( !std::isfinite ( mij ) ) { return false ; } // finite ?
+	  const T mji = mtrx ( j , i ) ;
+	  if ( !std::isfinite ( mji ) ) { return false ; } // finite ?
+	  if ( !equal ( mij , mji )   ) { return false ; } // symmetric? 
+	} 	
+      }
+      // ======================================================================
+      /// convert to symmetric matrix 
+      ROOT::Math::SMatrix<T,D,D,ROOT::Math::MatRepSym<T,D>> sym_mtrx ;
+      ROOT::Math::AssignSym::Evaluate ( sym_mtrx , mtrx ) ;
+      /// check for positive definitess  
+      const ROOT::Math::CholeskyDecomp<T,D> decomp ( sym_mtrx ) ;
+      return decomp.ok () ;      
+    }
+    // ========================================================================
+    /** Can this matrix be symmetric & positivee definite ?
+     *  - Diagonal elements are finite and positive
+     *  - Off-diagonal elements are finite
+     *  - have CholeskyDecomposition
+     */
+    template <class T, unsigned int D>
+    inline bool symmetric_positive_definite 
+    ( const ROOT::Math::SMatrix<T,D,D,ROOT::Math::MatRepSym<T,D>> & mtrx )
+    {
+      //
+      const Ostap::Math::Zero<T>     zero;
+      for ( unsigned int i = 0 ; i < D ; ++i )
+      {
+	/// diagonal elements must be positive! 
+	const T mii = mtrx ( i , i ) ;
+	if ( !std::isfinite ( mii ) || mii <= 0 || zero ( mii ) ) { return false ; }
+	for ( unsigned j = 0 ; j < i ; ++j )
+	{
+	  const T mjj = mtrx ( j , j ) ;
+	  const T mij = mtrx ( i , j ) ;
+	  /// off-diagonal elements can not be too large
+	  if  ( !std::isfinite ( mij ) || mij * mij > mii * mjj ) { return false ; }
+	}
+      }
+      /// check for positive definitess  
+      const ROOT::Math::CholeskyDecomp<T,D> decomp ( mtrx ) ;
+      return decomp.ok () ;      
+    }
+
+    // ========================================================================
+    
+    // ========================================================================
+    /** Can this matrix be a covariance matrix? 
+     *  - Diagonal elements are finite and positive
+     *  - Off-diagonal elements are finite and not too large 
+     *  - have CholeskyDecomposition
+     */
+    template <class T, unsigned int D1, unsigned int D2, class R>
+    inline bool covariance_matrix 
+    ( const ROOT::Math::SMatrix<T,D1,D2,R> & /* mtrx */ )
+    { return false ; } 
+
+    // ========================================================================
+    /** Can this matrix be a covariance matrix? 
+     *  - Diagonal elements are finite and positive
+     *  - Symmetrical 
+     *  - Off-diagonal elements are finite and not too large 
+     *  - have CholeskyDecomposition
+     */
+    template <class T, unsigned int D, class R>
+    inline bool covariance_matrix 
+    ( const ROOT::Math::SMatrix<T,D,D,R> & mtrx )
+    {
+      //
+      const Ostap::Math::Zero<T>     zero  ;
+      const Ostap::Math::Equal_To<T> equal ;
+      //
+      for ( unsigned int i = 0 ; i < D ; ++i )
+      {
+	/// diagonal elements must be positive! 
+	const T mii = mtrx ( i , i ) ;
+	if ( !std::isfinite ( mii ) || mii <= 0 || zero ( mii ) ) { return false ; } // positive ? 
+	for ( unsigned j = 0 ; j < i ; ++j )
+	{
+	  //
+	  const T mij = mtrx ( i , j ) ;
+	  if ( !std::isfinite ( mij ) ) { return false ; } // finite ?
+	  const T mji = mtrx ( j , i ) ;
+	  if ( !std::isfinite ( mji ) ) { return false ; } // finite ?
+	  //
+	  if ( !equal ( mij , mji )   ) { return false ; } // symmetric? 
+	  //
+	  const T mjj = mtrx ( j , j ) ;
+	  //
+	  /// off-diagonal elements can not be too large
+	  const T mm = mii * mjj ;
+	  if  ( mij * mij > mm ) { return false ; }        // too large 
+	  if  ( mji * mji > mm ) { return false ; }        // too large 
+	  // 
+	} 	
+      }
+      // ======================================================================
+      // convert to symmetric matrix 
+      ROOT::Math::SMatrix<T,D,D,ROOT::Math::MatRepSym<T,D>> sym_mtrx ;
+      ROOT::Math::AssignSym::Evaluate ( sym_mtrx , mtrx ) ;
+      /// check for positive definitess  
+      const ROOT::Math::CholeskyDecomp<T,D> decomp ( sym_mtrx ) ;
+      return decomp.ok () ;      
+    }
+    // ========================================================================
+    /** Can this matrix be a covariance matrix? 
+     *  - Diagonal elements are finite and positive
+     *  - Off-diagonal elements are finite and not too large 
+     *  - have CholeskyDecomposition
+     */
+    template <class T, unsigned int D>
+    inline bool covariance_matrix 
+    ( const ROOT::Math::SMatrix<T,D,D,ROOT::Math::MatRepSym<T,D>> & mtrx )
+    {
+      //
+      const Ostap::Math::Zero<T>     zero;
+      for ( unsigned int i = 0 ; i < D ; ++i )
+      {
+	/// diagonal elements must be positive! 
+	const T mii = mtrx ( i , i ) ;
+	if ( !std::isfinite ( mii ) || mii <= 0 || zero ( mii ) ) { return false ; }
+	for ( unsigned j = 0 ; j < i ; ++j )
+	{
+	  const T mjj = mtrx ( j , j ) ;
+	  const T mij = mtrx ( i , j ) ;
+	  /// off-diagonal elements can not be too large
+	  if  ( !std::isfinite ( mij ) || mij * mij > mii * mjj ) { return false ; }
+	}
+      }
+      /// check for positive definitess  
+      const ROOT::Math::CholeskyDecomp<T,D> decomp ( mtrx ) ;
+      return decomp.ok () ;      
+    }
+    // ========================================================================
   } //                                                    end of namespace Math
   // ==========================================================================
 } //                                               end of namespace Ostap::Math